move some logic to seperate files + renaming functions and pep8 styling
